backend/fasapi/requirements.txt
annotated-types==0.7.0
anyio==4.9.0
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
async-timeout==5.0.1
bcrypt==3.2.0
certifi==2025.4.26
cffi==1.17.1
click==8.2.0
colorama==0.4.6
ecdsa==0.19.1
fastapi==0.115.12
greenlet==3.2.2
grpcio==1.71.0
grpcio-tools==1.71.0
h11==0.16.0
idna==3.10
minio==7.2.15
passlib==1.7.4
protobuf==5.29.4
psycopg2==2.9.10
psycopg2-binary==2.9.10
pyasn1==0.4.8
pycparser==2.22
pycryptodome==3.22.0
pydantic==2.11.4
pydantic_core==2.33.2
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
redis==6.0.0
rsa==4.9.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2